test: consolidate check, tui, mcp, sqlite-vec, schema-inference, extraction, validation, facade, filters
Wave 2 of test suite consolidation. Merges 9 satellite test files into their
natural primaries:

- test_check_command_integration → test_check
- test_facade_helpers_coverage → test_facade
- test_filters_schemas_coverage → test_filters
- test_mcp_server_integration → test_mcp_server
- test_schema_drift → test_schema_inference
- test_schema_inference_coverage → test_schema_inference
- test_sqlite_vec_operations → test_sqlite_vec
- test_tui_comprehensive → test_tui
- test_unified_extraction_edge_cases → test_unified_extraction

Also consolidates overlapping validation tests. Zero test regressions.
