+++ tests/test_facade.py
@@ -3,11 +3,28 @@
 from __future__ import annotations
 
 import json
+import os
+from datetime import datetime, timezone
+from pathlib import Path
+from unittest.mock import MagicMock, Mock, patch
 
 import pytest
 
 from polylogue import Polylogue
-from polylogue.config import Source
+from polylogue.cli.helpers import (
+    latest_render_path,
+    load_last_source,
+    maybe_prompt_sources,
+    print_summary,
+    save_last_source,
+)
+from polylogue.cli.types import AppEnv
+from polylogue.config import Config, Source
+from polylogue.facade import ArchiveStats
+from polylogue.lib.models import Conversation, Message
+from polylogue.lib.messages import MessageCollection
+from polylogue.storage.backends.sqlite import SQLiteBackend
+from polylogue.storage.store import ConversationRecord, MessageRecord
 
 
 @pytest.fixture
@@ -447,3 +464,945 @@ class TestPolylogueEdgeCases:
         result = archive.ingest_file(nonexistent)
         assert result.counts["conversations"] == 0
         assert result.counts["messages"] == 0
+
+
+# MERGED FROM test_facade_helpers_coverage.py
+
+# ============================================================================
+# ARCHIVE STATS TESTS
+# ============================================================================
+
+
+class TestArchiveStatsCreation:
+    """Test ArchiveStats instantiation and attributes."""
+
+    def test_archive_stats_init_basic(self):
+        """Test basic ArchiveStats initialization."""
+        stats = ArchiveStats(
+            conversation_count=10,
+            message_count=50,
+            word_count=1000,
+            providers={"claude": 7, "chatgpt": 3},
+            tags={"test": 2, "work": 3},
+            last_sync=None,
+            recent=[],
+        )
+        assert stats.conversation_count == 10
+        assert stats.message_count == 50
+        assert stats.word_count == 1000
+        assert stats.providers == {"claude": 7, "chatgpt": 3}
+        assert stats.tags == {"test": 2, "work": 3}
+        assert stats.last_sync is None
+        assert stats.recent == []
+
+    def test_archive_stats_with_timestamp(self):
+        """Test ArchiveStats with last_sync timestamp."""
+        timestamp = "2025-01-15T12:30:45Z"
+        stats = ArchiveStats(
+            conversation_count=5,
+            message_count=25,
+            word_count=500,
+            providers={"claude": 5},
+            tags={},
+            last_sync=timestamp,
+            recent=[],
+        )
+        assert stats.last_sync == timestamp
+
+    def test_archive_stats_with_recent_conversations(self):
+        """Test ArchiveStats with recent conversations."""
+        recent_msgs = [
+            Message(id="m1", role="user", text="Hello", timestamp=datetime.now(tz=timezone.utc)),
+        ]
+        recent_conv = Conversation(
+            id="conv1",
+            provider="claude",
+            messages=MessageCollection(messages=recent_msgs),
+        )
+        stats = ArchiveStats(
+            conversation_count=1,
+            message_count=1,
+            word_count=10,
+            providers={"claude": 1},
+            tags={},
+            last_sync=None,
+            recent=[recent_conv],
+        )
+        assert len(stats.recent) == 1
+        assert stats.recent[0].id == "conv1"
+
+    def test_archive_stats_repr(self):
+        """Test ArchiveStats __repr__ includes key information."""
+        stats = ArchiveStats(
+            conversation_count=10,
+            message_count=50,
+            word_count=1000,
+            providers={"claude": 7, "chatgpt": 3},
+            tags={"test": 2},
+            last_sync=None,
+            recent=[],
+        )
+        repr_str = repr(stats)
+        assert "10" in repr_str
+        assert "50" in repr_str
+        assert "ArchiveStats" in repr_str
+
+    def test_archive_stats_empty_providers(self):
+        """Test ArchiveStats with empty providers."""
+        stats = ArchiveStats(
+            conversation_count=0,
+            message_count=0,
+            word_count=0,
+            providers={},
+            tags={},
+            last_sync=None,
+            recent=[],
+        )
+        assert stats.providers == {}
+        assert len(stats.recent) == 0
+
+
+# ============================================================================
+# POLYLOGUE INITIALIZATION TESTS
+# ============================================================================
+
+
+class TestPolylogueInit:
+    """Test Polylogue initialization with various configurations."""
+
+    def test_polylogue_init_with_memory_db(self, tmp_path):
+        """Test Polylogue initialization with in-memory database."""
+        archive = Polylogue(archive_root=tmp_path, db_path=":memory:")
+        assert archive is not None
+        assert archive.archive_root == tmp_path
+
+    def test_polylogue_init_with_file_db(self, tmp_path):
+        """Test Polylogue initialization with file-based database."""
+        db_path = tmp_path / "test.db"
+        archive = Polylogue(archive_root=tmp_path, db_path=db_path)
+        assert archive.config is not None
+        assert archive.repository is not None
+
+    def test_polylogue_config_property(self, tmp_path):
+        """Test Polylogue.config property returns Config object."""
+        archive = Polylogue(archive_root=tmp_path, db_path=":memory:")
+        cfg = archive.config
+        assert type(cfg).__name__ == "Config"
+        assert cfg.archive_root is not None
+
+    def test_polylogue_repository_property(self, tmp_path):
+        """Test Polylogue.repository property returns repository."""
+        archive = Polylogue(archive_root=tmp_path, db_path=":memory:")
+        repo = archive.repository
+        assert repo is not None
+
+    def test_polylogue_archive_root_property(self, tmp_path):
+        """Test Polylogue.archive_root property."""
+        archive = Polylogue(archive_root=tmp_path, db_path=":memory:")
+        assert archive.archive_root == tmp_path
+
+    def test_polylogue_repr(self, tmp_path):
+        """Test Polylogue __repr__ method."""
+        archive = Polylogue(archive_root=tmp_path, db_path=":memory:")
+        repr_str = repr(archive)
+        assert "Polylogue" in repr_str
+        assert "archive_root" in repr_str
+
+
+# ============================================================================
+# POLYLOGUE CONVERSATION RETRIEVAL TESTS
+# ============================================================================
+
+
+class TestPolylogueGetConversation:
+    """Test getting single conversations."""
+
+    def test_get_conversation_empty_db(self, tmp_path):
+        """Test getting conversation from empty database."""
+        db_path = tmp_path / "test.db"
+        archive = Polylogue(archive_root=tmp_path, db_path=db_path)
+        conv = archive.get_conversation("nonexistent_id")
+        assert conv is None
+
+    def test_get_conversation_with_seed_data(self, tmp_path):
+        """Test retrieving a conversation after adding data."""
+        db_path = tmp_path / "test.db"
+        archive = Polylogue(archive_root=tmp_path, db_path=db_path)
+        backend = archive.repository.backend
+
+        # Create and save a conversation
+        conv_record = ConversationRecord(
+            conversation_id="conv-1",
+            provider_name="claude",
+            provider_conversation_id="provider-1",
+            title="Test Conversation",
+            created_at="2025-01-01T00:00:00Z",
+            updated_at="2025-01-01T00:00:00Z",
+            content_hash="hash-1",
+        )
+        backend.save_conversation(conv_record)
+
+        # Save messages
+        msg_records = [
+            MessageRecord(
+                message_id="msg-1",
+                conversation_id="conv-1",
+                role="user",
+                text="Hello",
+                timestamp="2025-01-01T00:00:00Z",
+                content_hash="msg-hash-1",
+            ),
+            MessageRecord(
+                message_id="msg-2",
+                conversation_id="conv-1",
+                role="assistant",
+                text="Hi there",
+                timestamp="2025-01-01T00:01:00Z",
+                content_hash="msg-hash-2",
+            ),
+        ]
+        backend.save_messages(msg_records)
+
+        # Retrieve by ID
+        conv = archive.get_conversation("conv-1")
+        assert conv is not None
+        assert conv.id == "conv-1"
+        assert conv.title == "Test Conversation"
+
+
+class TestPolylogueGetConversations:
+    """Test batch conversation retrieval."""
+
+    def test_get_conversations_empty_list(self, tmp_path):
+        """Test get_conversations with empty ID list."""
+        db_path = tmp_path / "test.db"
+        archive = Polylogue(archive_root=tmp_path, db_path=db_path)
+        convs = archive.get_conversations([])
+        assert convs == []
+
+    def test_get_conversations_batch(self, tmp_path):
+        """Test batch retrieval of multiple conversations."""
+        db_path = tmp_path / "test.db"
+        archive = Polylogue(archive_root=tmp_path, db_path=db_path)
+        backend = archive.repository.backend
+
+        # Create multiple conversations
+        for i in range(3):
+            conv_record = ConversationRecord(
+                conversation_id=f"conv-{i}",
+                provider_name="claude",
+                provider_conversation_id=f"provider-{i}",
+                title=f"Conversation {i}",
+                created_at="2025-01-01T00:00:00Z",
+                updated_at="2025-01-01T00:00:00Z",
+                content_hash=f"hash-{i}",
+            )
+            backend.save_conversation(conv_record)
+
+        # Retrieve batch
+        ids = ["conv-0", "conv-1", "conv-2"]
+        convs = archive.get_conversations(ids)
+        assert len(convs) == 3
+        assert all(c.id in ids for c in convs)
+
+    def test_get_conversations_partial_match(self, tmp_path):
+        """Test batch retrieval with some missing IDs."""
+        db_path = tmp_path / "test.db"
+        archive = Polylogue(archive_root=tmp_path, db_path=db_path)
+        backend = archive.repository.backend
+
+        # Create only conv-1
+        conv_record = ConversationRecord(
+            conversation_id="conv-1",
+            provider_name="claude",
+            provider_conversation_id="provider-1",
+            title="Conversation 1",
+            created_at="2025-01-01T00:00:00Z",
+            updated_at="2025-01-01T00:00:00Z",
+            content_hash="hash-1",
+        )
+        backend.save_conversation(conv_record)
+
+        # Request multiple IDs but only conv-1 exists
+        ids = ["conv-1", "conv-999"]
+        convs = archive.get_conversations(ids)
+        assert len(convs) == 1
+        assert convs[0].id == "conv-1"
+
+
+# ============================================================================
+# POLYLOGUE LIST CONVERSATIONS TESTS
+# ============================================================================
+
+
+class TestPolylogueListConversations:
+    """Test listing conversations with various filters."""
+
+    def test_list_conversations_empty_db(self, tmp_path):
+        """Test listing from empty database."""
+        db_path = tmp_path / "test.db"
+        archive = Polylogue(archive_root=tmp_path, db_path=db_path)
+        convs = archive.list_conversations()
+        assert convs == []
+
+    def test_list_conversations_all(self, tmp_path):
+        """Test listing all conversations."""
+        db_path = tmp_path / "test.db"
+        archive = Polylogue(archive_root=tmp_path, db_path=db_path)
+        backend = archive.repository.backend
+
+        for i in range(3):
+            conv_record = ConversationRecord(
+                conversation_id=f"conv-{i}",
+                provider_name="claude" if i < 2 else "chatgpt",
+                provider_conversation_id=f"provider-{i}",
+                title=f"Conv {i}",
+                created_at="2025-01-01T00:00:00Z",
+                updated_at="2025-01-01T00:00:00Z",
+                content_hash=f"hash-{i}",
+            )
+            backend.save_conversation(conv_record)
+
+        convs = archive.list_conversations()
+        assert len(convs) == 3
+
+    def test_list_conversations_filter_by_provider(self, tmp_path):
+        """Test listing with provider filter."""
+        db_path = tmp_path / "test.db"
+        archive = Polylogue(archive_root=tmp_path, db_path=db_path)
+        backend = archive.repository.backend
+
+        for i in range(2):
+            backend.save_conversation(ConversationRecord(
+                conversation_id=f"claude-{i}",
+                provider_name="claude",
+                provider_conversation_id=f"p-{i}",
+                title=f"Claude {i}",
+                created_at="2025-01-01T00:00:00Z",
+                updated_at="2025-01-01T00:00:00Z",
+                content_hash=f"h-{i}",
+            ))
+        for i in range(2, 4):
+            backend.save_conversation(ConversationRecord(
+                conversation_id=f"chatgpt-{i}",
+                provider_name="chatgpt",
+                provider_conversation_id=f"p-{i}",
+                title=f"ChatGPT {i}",
+                created_at="2025-01-01T00:00:00Z",
+                updated_at="2025-01-01T00:00:00Z",
+                content_hash=f"h-{i}",
+            ))
+
+        claude_convs = archive.list_conversations(provider="claude")
+        assert len(claude_convs) == 2
+        assert all(c.provider == "claude" for c in claude_convs)
+
+    def test_list_conversations_with_limit(self, tmp_path):
+        """Test listing with limit."""
+        db_path = tmp_path / "test.db"
+        archive = Polylogue(archive_root=tmp_path, db_path=db_path)
+        backend = archive.repository.backend
+
+        for i in range(5):
+            backend.save_conversation(ConversationRecord(
+                conversation_id=f"conv-{i}",
+                provider_name="claude",
+                provider_conversation_id=f"p-{i}",
+                title=f"Conv {i}",
+                created_at="2025-01-01T00:00:00Z",
+                updated_at="2025-01-01T00:00:00Z",
+                content_hash=f"h-{i}",
+            ))
+
+        convs = archive.list_conversations(limit=2)
+        assert len(convs) == 2
+
+    def test_list_conversations_filter_by_source(self, tmp_path):
+        """Test listing with source filter."""
+        db_path = tmp_path / "test.db"
+        archive = Polylogue(archive_root=tmp_path, db_path=db_path)
+        backend = archive.repository.backend
+
+        # Create conversations with different sources in provider_meta
+        backend.save_conversation(ConversationRecord(
+            conversation_id="conv-1",
+            provider_name="claude",
+            provider_conversation_id="p-1",
+            title="Conv 1",
+            created_at="2025-01-01T00:00:00Z",
+            updated_at="2025-01-01T00:00:00Z",
+            content_hash="h-1",
+            provider_meta={"source": "source-a"},
+        ))
+        backend.save_conversation(ConversationRecord(
+            conversation_id="conv-2",
+            provider_name="claude",
+            provider_conversation_id="p-2",
+            title="Conv 2",
+            created_at="2025-01-01T00:00:00Z",
+            updated_at="2025-01-01T00:00:00Z",
+            content_hash="h-2",
+            provider_meta={"source": "source-b"},
+        ))
+
+        convs = archive.list_conversations(source="source-a")
+        assert len(convs) == 1
+        assert convs[0].id == "conv-1"
+
+    def test_list_conversations_source_and_limit(self, tmp_path):
+        """Test listing with both source and limit filters."""
+        db_path = tmp_path / "test.db"
+        archive = Polylogue(archive_root=tmp_path, db_path=db_path)
+        backend = archive.repository.backend
+
+        # Create multiple conversations with same source
+        for i in range(5):
+            backend.save_conversation(ConversationRecord(
+                conversation_id=f"conv-{i}",
+                provider_name="claude",
+                provider_conversation_id=f"p-{i}",
+                title=f"Conv {i}",
+                created_at="2025-01-01T00:00:00Z",
+                updated_at="2025-01-01T00:00:00Z",
+                content_hash=f"h-{i}",
+                provider_meta={"source": "inbox"},
+            ))
+
+        convs = archive.list_conversations(source="inbox", limit=3)
+        assert len(convs) == 3
+
+
+# ============================================================================
+# POLYLOGUE FILTER AND CONTEXT MANAGER TESTS
+# ============================================================================
+
+
+class TestPolylogueFilter:
+    """Test filter builder creation."""
+
+    def test_filter_returns_conversation_filter(self, tmp_path):
+        """Test that filter() returns a ConversationFilter."""
+        db_path = tmp_path / "test.db"
+        archive = Polylogue(archive_root=tmp_path, db_path=db_path)
+        filter_builder = archive.filter()
+        assert filter_builder is not None
+        # ConversationFilter has provider method
+        assert hasattr(filter_builder, "provider")
+
+    def test_filter_chaining(self, tmp_path):
+        """Test filter method chaining."""
+        db_path = tmp_path / "test.db"
+        archive = Polylogue(archive_root=tmp_path, db_path=db_path)
+        filter_builder = archive.filter()
+        # Should support chaining
+        result = filter_builder.provider("claude")
+        assert result is not None
+
+
+class TestPolylogueContextManager:
+    """Test Polylogue as context manager."""
+
+    def test_context_manager_enter_returns_self(self, tmp_path):
+        """Test __enter__ returns self."""
+        db_path = tmp_path / "test.db"
+        archive = Polylogue(archive_root=tmp_path, db_path=db_path)
+        with archive as ctx:
+            assert ctx is archive
+
+    def test_context_manager_exit_calls_close(self, tmp_path):
+        """Test __exit__ calls close."""
+        db_path = tmp_path / "test.db"
+        archive = Polylogue(archive_root=tmp_path, db_path=db_path)
+        # Mock close to verify it's called
+        archive.close = MagicMock()
+        with archive:
+            pass
+        archive.close.assert_called_once()
+
+    def test_context_manager_with_exception(self, tmp_path):
+        """Test context manager properly closes on exception."""
+        db_path = tmp_path / "test.db"
+        archive = Polylogue(archive_root=tmp_path, db_path=db_path)
+        archive.close = MagicMock()
+
+        try:
+            with archive:
+                raise ValueError("Test error")
+        except ValueError:
+            pass
+
+        archive.close.assert_called_once()
+
+    def test_close_method(self, tmp_path):
+        """Test close() method."""
+        db_path = tmp_path / "test.db"
+        archive = Polylogue(archive_root=tmp_path, db_path=db_path)
+        # Should not raise
+        archive.close()
+
+
+# ============================================================================
+# POLYLOGUE REBUILD INDEX AND STATS TESTS
+# ============================================================================
+
+
+class TestPolylogueRebuildIndex:
+    """Test index rebuilding."""
+
+    def test_rebuild_index_lazy_init(self, tmp_path):
+        """Test that rebuild_index lazy-initializes the service."""
+        db_path = tmp_path / "test.db"
+        archive = Polylogue(archive_root=tmp_path, db_path=db_path)
+        assert archive._indexing_service is None
+
+        # Just verify the method exists and can be called
+        assert hasattr(archive, "rebuild_index")
+        assert callable(archive.rebuild_index)
+
+
+class TestPolylogueStats:
+    """Test statistics generation."""
+
+    def test_stats_empty_db(self, tmp_path):
+        """Test stats() on empty database."""
+        db_path = tmp_path / "test.db"
+        archive = Polylogue(archive_root=tmp_path, db_path=db_path)
+        stats = archive.stats()
+
+        assert isinstance(stats, ArchiveStats)
+        assert stats.conversation_count == 0
+        assert stats.message_count == 0
+        assert stats.word_count == 0
+        assert stats.providers == {}
+        assert stats.tags == {}
+
+    def test_stats_with_conversations(self, tmp_path):
+        """Test stats() with conversations in database."""
+        db_path = tmp_path / "test.db"
+        archive = Polylogue(archive_root=tmp_path, db_path=db_path)
+        backend = archive.repository.backend
+
+        # Create conversations with different providers
+        for i in range(2):
+            conv_record = ConversationRecord(
+                conversation_id=f"claude-{i}",
+                provider_name="claude",
+                provider_conversation_id=f"p-{i}",
+                title=f"Claude Conv {i}",
+                created_at="2025-01-01T00:00:00Z",
+                updated_at="2025-01-01T00:00:00Z",
+                content_hash=f"h-{i}",
+            )
+            backend.save_conversation(conv_record)
+
+            # Add messages
+            msg_records = [
+                MessageRecord(
+                    message_id=f"msg-{i}-0",
+                    conversation_id=f"claude-{i}",
+                    role="user",
+                    text="Hello world",
+                    timestamp="2025-01-01T00:00:00Z",
+                    content_hash=f"mh-{i}-0",
+                ),
+                MessageRecord(
+                    message_id=f"msg-{i}-1",
+                    conversation_id=f"claude-{i}",
+                    role="assistant",
+                    text="Hi there friend",
+                    timestamp="2025-01-01T00:01:00Z",
+                    content_hash=f"mh-{i}-1",
+                ),
+            ]
+            backend.save_messages(msg_records)
+
+        # Add ChatGPT conversation
+        conv_record = ConversationRecord(
+            conversation_id="chatgpt-0",
+            provider_name="chatgpt",
+            provider_conversation_id="p-chatgpt-0",
+            title="ChatGPT Conv",
+            created_at="2025-01-01T00:00:00Z",
+            updated_at="2025-01-01T00:00:00Z",
+            content_hash="h-chatgpt",
+        )
+        backend.save_conversation(conv_record)
+        msg_records = [
+            MessageRecord(
+                message_id="msg-chatgpt-0",
+                conversation_id="chatgpt-0",
+                role="user",
+                text="Test",
+                timestamp="2025-01-01T00:00:00Z",
+                content_hash="mh-chatgpt",
+            ),
+        ]
+        backend.save_messages(msg_records)
+
+        stats = archive.stats()
+        assert stats.conversation_count == 3
+        assert stats.message_count == 5
+        assert "claude" in stats.providers
+        assert "chatgpt" in stats.providers
+        assert stats.providers["claude"] == 2
+        assert stats.providers["chatgpt"] == 1
+
+    def test_stats_recent_conversations(self, tmp_path):
+        """Test that stats includes recent conversations."""
+        db_path = tmp_path / "test.db"
+        archive = Polylogue(archive_root=tmp_path, db_path=db_path)
+        backend = archive.repository.backend
+
+        # Create a single conversation
+        conv_record = ConversationRecord(
+            conversation_id="conv-1",
+            provider_name="claude",
+            provider_conversation_id="p-1",
+            title="Test Conv",
+            created_at="2025-01-01T00:00:00Z",
+            updated_at="2025-01-15T12:00:00Z",
+            content_hash="h-1",
+        )
+        backend.save_conversation(conv_record)
+
+        stats = archive.stats()
+        assert len(stats.recent) == 1
+        assert stats.recent[0].id == "conv-1"
+
+
+# ============================================================================
+# CLI HELPERS TESTS: SOURCE STATE MANAGEMENT
+# ============================================================================
+
+
+class TestLoadSaveLastSource:
+    """Test source state persistence."""
+
+    def test_save_last_source(self, tmp_path, monkeypatch):
+        """Test saving last source."""
+        state_dir = tmp_path / "state"
+        monkeypatch.setenv("XDG_STATE_HOME", str(state_dir))
+
+        save_last_source("my-source")
+
+        path = state_dir / "polylogue" / "last-source.json"
+        assert path.exists()
+        data = json.loads(path.read_text(encoding="utf-8"))
+        assert data["source"] == "my-source"
+
+    def test_load_last_source_exists(self, tmp_path, monkeypatch):
+        """Test loading existing last source."""
+        state_dir = tmp_path / "state"
+        state_dir.mkdir(parents=True, exist_ok=True)
+        monkeypatch.setenv("XDG_STATE_HOME", str(state_dir))
+
+        # Create state file
+        source_file = state_dir / "polylogue" / "last-source.json"
+        source_file.parent.mkdir(parents=True, exist_ok=True)
+        source_file.write_text(json.dumps({"source": "test-source"}), encoding="utf-8")
+
+        result = load_last_source()
+        assert result == "test-source"
+
+    def test_load_last_source_not_exists(self, tmp_path, monkeypatch):
+        """Test loading when file doesn't exist."""
+        state_dir = tmp_path / "state"
+        monkeypatch.setenv("XDG_STATE_HOME", str(state_dir))
+
+        result = load_last_source()
+        assert result is None
+
+    def test_load_last_source_invalid_json(self, tmp_path, monkeypatch):
+        """Test loading with invalid JSON."""
+        state_dir = tmp_path / "state"
+        state_dir.mkdir(parents=True, exist_ok=True)
+        monkeypatch.setenv("XDG_STATE_HOME", str(state_dir))
+
+        source_file = state_dir / "polylogue" / "last-source.json"
+        source_file.parent.mkdir(parents=True, exist_ok=True)
+        source_file.write_text("invalid json", encoding="utf-8")
+
+        result = load_last_source()
+        assert result is None
+
+    def test_load_last_source_missing_field(self, tmp_path, monkeypatch):
+        """Test loading when 'source' field is missing."""
+        state_dir = tmp_path / "state"
+        state_dir.mkdir(parents=True, exist_ok=True)
+        monkeypatch.setenv("XDG_STATE_HOME", str(state_dir))
+
+        source_file = state_dir / "polylogue" / "last-source.json"
+        source_file.parent.mkdir(parents=True, exist_ok=True)
+        source_file.write_text(json.dumps({"other": "value"}), encoding="utf-8")
+
+        result = load_last_source()
+        assert result is None
+
+
+# ============================================================================
+# CLI HELPERS TESTS: maybe_prompt_sources
+# ============================================================================
+
+
+class TestMaybePromptSources:
+    """Test interactive source selection."""
+
+    def test_maybe_prompt_already_selected(self):
+        """Test when sources are already selected."""
+        ui_mock = MagicMock()
+        env = AppEnv(ui=ui_mock)
+        config = Config(
+            archive_root=Path("/tmp"),
+            render_root=Path("/tmp/render"),
+            sources=[Source(name="source-a", path=Path("/a")), Source(name="source-b", path=Path("/b"))],
+        )
+        selected = ["source-a"]
+
+        result = maybe_prompt_sources(env, config, selected, "test-cmd")
+        assert result == ["source-a"]
+        ui_mock.choose.assert_not_called()
+
+    def test_maybe_prompt_plain_mode(self):
+        """Test plain mode passes through."""
+        ui_mock = MagicMock()
+        ui_mock.plain = True
+        env = AppEnv(ui=ui_mock)
+        config = Config(
+            archive_root=Path("/tmp"),
+            render_root=Path("/tmp/render"),
+            sources=[Source(name="source-a", path=Path("/a")), Source(name="source-b", path=Path("/b"))],
+        )
+
+        result = maybe_prompt_sources(env, config, None, "test-cmd")
+        assert result is None
+        ui_mock.choose.assert_not_called()
+
+    def test_maybe_prompt_single_source(self):
+        """Test with single source (no prompt needed)."""
+        ui_mock = MagicMock()
+        ui_mock.plain = False
+        env = AppEnv(ui=ui_mock)
+        config = Config(
+            archive_root=Path("/tmp"),
+            render_root=Path("/tmp/render"),
+            sources=[Source(name="only-source", path=Path("/a"))],
+        )
+
+        result = maybe_prompt_sources(env, config, None, "test-cmd")
+        assert result is None
+        ui_mock.choose.assert_not_called()
+
+    def test_maybe_prompt_multiple_sources(self):
+        """Test interactive prompt with multiple sources."""
+        ui_mock = MagicMock()
+        ui_mock.plain = False
+        ui_mock.choose.return_value = "source-b"
+        env = AppEnv(ui=ui_mock)
+
+        config = Config(
+            archive_root=Path("/tmp"),
+            render_root=Path("/tmp/render"),
+            sources=[Source(name="source-a", path=Path("/a")), Source(name="source-b", path=Path("/b"))],
+        )
+
+        with patch("polylogue.cli.helpers.load_last_source", return_value=None):
+            with patch("polylogue.cli.helpers.save_last_source"):
+                result = maybe_prompt_sources(env, config, None, "test-cmd")
+
+        assert result == ["source-b"]
+        ui_mock.choose.assert_called_once()
+
+    def test_maybe_prompt_all_selected(self):
+        """Test when 'all' is chosen."""
+        ui_mock = MagicMock()
+        ui_mock.plain = False
+        ui_mock.choose.return_value = "all"
+        env = AppEnv(ui=ui_mock)
+
+        config = Config(
+            archive_root=Path("/tmp"),
+            render_root=Path("/tmp/render"),
+            sources=[Source(name="source-a", path=Path("/a")), Source(name="source-b", path=Path("/b"))],
+        )
+
+        with patch("polylogue.cli.helpers.load_last_source", return_value=None):
+            with patch("polylogue.cli.helpers.save_last_source"):
+                result = maybe_prompt_sources(env, config, None, "test-cmd")
+
+        assert result is None
+
+    def test_maybe_prompt_restores_last_choice(self):
+        """Test that last choice is restored to top of options."""
+        ui_mock = MagicMock()
+        ui_mock.plain = False
+        ui_mock.choose.return_value = "source-b"
+        env = AppEnv(ui=ui_mock)
+
+        config = Config(
+            archive_root=Path("/tmp"),
+            render_root=Path("/tmp/render"),
+            sources=[Source(name="source-a", path=Path("/a")), Source(name="source-b", path=Path("/b"))],
+        )
+
+        with patch("polylogue.cli.helpers.load_last_source", return_value="source-b"):
+            with patch("polylogue.cli.helpers.save_last_source"):
+                maybe_prompt_sources(env, config, None, "test-cmd")
+
+        # Verify choose was called with options list that has source-b at top
+        call_args = ui_mock.choose.call_args
+        options = call_args[0][1]
+        assert options[0] == "source-b"
+
+    def test_maybe_prompt_no_choice_fails(self):
+        """Test that not selecting a source fails."""
+        ui_mock = MagicMock()
+        ui_mock.plain = False
+        ui_mock.choose.return_value = None
+        env = AppEnv(ui=ui_mock)
+
+        config = Config(
+            archive_root=Path("/tmp"),
+            render_root=Path("/tmp/render"),
+            sources=[Source(name="source-a", path=Path("/a")), Source(name="source-b", path=Path("/b"))],
+        )
+
+        with patch("polylogue.cli.helpers.load_last_source", return_value=None):
+            with pytest.raises(SystemExit):
+                maybe_prompt_sources(env, config, None, "test-cmd")
+
+
+# ============================================================================
+# CLI HELPERS TESTS: latest_render_path
+# ============================================================================
+
+
+class TestLatestRenderPath:
+    """Test finding latest rendered conversation."""
+
+    def test_latest_render_path_no_root(self, tmp_path):
+        """Test when render root doesn't exist."""
+        nonexistent = tmp_path / "nonexistent"
+        result = latest_render_path(nonexistent)
+        assert result is None
+
+    def test_latest_render_path_empty_root(self, tmp_path):
+        """Test when render root is empty."""
+        result = latest_render_path(tmp_path)
+        assert result is None
+
+    def test_latest_render_path_md_files(self, tmp_path):
+        """Test finding latest .md file."""
+        import time
+
+        render_dir = tmp_path / "render"
+        render_dir.mkdir()
+
+        # Create nested structure with .md files
+        conv_dir = render_dir / "some_conv"
+        conv_dir.mkdir()
+
+        # Create multiple .md files with different timestamps
+        md1 = conv_dir / "conversation.md"
+        md1.write_text("old")
+        os.utime(md1, (100, 100))  # Set old timestamp
+
+        md2 = render_dir / "conversation.md"
+        md2.write_text("new")
+        time.sleep(0.01)  # Ensure timestamp difference
+        os.utime(md2, (200, 200))  # Set newer timestamp
+
+        result = latest_render_path(render_dir)
+        assert result is not None
+        assert result.name == "conversation.md"
+
+    def test_latest_render_path_html_files(self, tmp_path):
+        """Test finding latest .html file."""
+        render_dir = tmp_path / "render"
+        render_dir.mkdir()
+
+        html_file = render_dir / "conversation.html"
+        html_file.write_text("<html></html>")
+
+        result = latest_render_path(render_dir)
+        assert result is not None
+        assert result.suffix == ".html"
+
+    def test_latest_render_path_handles_deleted_files(self, tmp_path):
+        """Test handling of files deleted between rglob and stat."""
+        render_dir = tmp_path / "render"
+        render_dir.mkdir()
+
+        # Create a file
+        md_file = render_dir / "conversation.md"
+        md_file.write_text("test")
+
+        # Patch rglob to return a path that will be deleted before stat
+        def mock_rglob(self, pattern):
+            # Return both the real file and a non-existent one
+            return [md_file, tmp_path / "deleted.md"]
+
+        with patch.object(Path, "rglob", mock_rglob):
+            result = latest_render_path(render_dir)
+            # Should handle the deleted file gracefully
+            assert result is not None or result is None  # Either is acceptable
+
+    def test_latest_render_path_mixed_formats(self, tmp_path):
+        """Test with both .md and .html files."""
+        render_dir = tmp_path / "render"
+        render_dir.mkdir()
+
+        md_file = render_dir / "conversation.md"
+        md_file.write_text("markdown")
+        os.utime(md_file, (100, 100))
+
+        html_file = render_dir / "conversation.html"
+        html_file.write_text("<html></html>")
+        os.utime(html_file, (200, 200))
+
+        result = latest_render_path(render_dir)
+        assert result is not None
+        # Should return the newer one
+        assert result.name in ["conversation.md", "conversation.html"]
+
+
+# ============================================================================
+# POLYLOGUE INGESTION TESTS
+# ============================================================================
+
+
+class TestPolylogueIngestFile:
+    """Test file ingestion."""
+
+    def test_ingest_file_method_exists(self, tmp_path):
+        """Test that ingest_file method exists and is callable."""
+        db_path = tmp_path / "test.db"
+        archive = Polylogue(archive_root=tmp_path, db_path=db_path)
+        assert hasattr(archive, "ingest_file")
+        assert callable(archive.ingest_file)
+
+    def test_ingest_file_with_nonexistent_file(self, tmp_path):
+        """Test ingest_file with non-existent file."""
+        db_path = tmp_path / "test.db"
+        archive = Polylogue(archive_root=tmp_path, db_path=db_path)
+
+        nonexistent = tmp_path / "does_not_exist.json"
+        # Ingestion should handle gracefully (empty result or exception)
+        # Most implementations skip missing files silently
+        try:
+            result = archive.ingest_file(nonexistent)
+            # If no exception, verify result structure
+            assert hasattr(result, "counts")
+        except Exception:
+            # Some implementations raise exceptions for missing files
+            pass
+
+
+class TestPolylogueIngestSources:
+    """Test sources ingestion."""
+
+    def test_ingest_sources_method_exists(self, tmp_path):
+        """Test that ingest_sources method exists and is callable."""
+        db_path = tmp_path / "test.db"
+        archive = Polylogue(archive_root=tmp_path, db_path=db_path)
+        assert hasattr(archive, "ingest_sources")
+        assert callable(archive.ingest_sources)
