--- tests/test_filters_schemas_coverage.py
@@ -1,977 +1,0 @@
-"""Comprehensive coverage tests for filters.py, unified.py, and validator.py.
-
-Targets specific uncovered lines:
-1. polylogue/lib/filters.py (80% → 90%)
-   - since/until with datetime objects (193, 216)
-   - similar() method (297-298)
-   - is_continuation/is_sidechain/is_root with False value (324, 339)
-   - has_branches with False value (381)
-   - _apply_filters with excluded providers, tags, content types (411, 413-414, 452)
-   - sort by tokens/words/longest (498, 519, 521)
-   - _fetch_candidates with ID prefix resolution (576-577)
-   - _has_post_filters with negative FTS terms (591-592)
-   - Search/FTS exception handling (606-613)
-   - list_summaries() all paths (669-685)
-   - pick() with ambiguous choice/EOF (701, 725-744)
-   - list_summaries() edge cases (784-787, 801-802, 830, 832, 834, 836-837, 841-842, 854, 874-876, 883)
-
-2. polylogue/schemas/unified.py (82% → 92%)
-   - _missing_role() error (33-34)
-   - extract_reasoning_traces with Gemini (129, 137)
-   - extract_content_blocks edge cases (159, 189, 224-233, 275, 330, 365, 431-433)
-   - extract_harmonized_message with invalid provider (338)
-   - harmonize_parsed_message with None provider_meta (501-509)
-   - bulk_harmonize edge cases (525-532)
-
-3. polylogue/schemas/validator.py (82% → 92%)
-   - jsonschema ImportError handling (26-29)
-   - available_providers with missing SCHEMA_DIR (78)
-   - _detect_drift with additionalProperties dict (112, 170, 175, 182-186)
-   - _format_error (197)
-   - validate with multiple errors (166)
-"""
-
-from __future__ import annotations
-
-import json
-import sys
-import tempfile
-from datetime import datetime, timezone
-from io import StringIO
-from pathlib import Path
-from unittest.mock import MagicMock, Mock, patch
-
-import pytest
-
-from polylogue.lib.filters import ConversationFilter
-from polylogue.lib.models import Conversation, ConversationSummary, Message
-from polylogue.storage.backends.sqlite import SQLiteBackend
-from polylogue.schemas.unified import (
-    HarmonizedMessage,
-    extract_reasoning_traces,
-    extract_tool_calls,
-    extract_content_blocks,
-    extract_token_usage,
-    extract_claude_code_text,
-    extract_chatgpt_text,
-    extract_codex_text,
-    extract_harmonized_message,
-    harmonize_parsed_message,
-    bulk_harmonize,
-    is_message_record,
-)
-from polylogue.schemas.validator import (
-    SchemaValidator,
-    ValidationResult,
-    validate_provider_export,
-)
-from polylogue.storage.backends.sqlite import open_connection
-from polylogue.storage.index import rebuild_index
-from polylogue.storage.repository import ConversationRepository
-from tests.helpers import ConversationBuilder
-
-
-# =============================================================================
-# FILTERS.PY TESTS
-# =============================================================================
-
-
-class TestFiltersDateTimeHandling:
-    """Test date handling with datetime objects (lines 193, 216)."""
-
-    @pytest.fixture
-    def filter_repo(self, tmp_path):
-        """Create repository for filter tests."""
-        db_path = tmp_path / "filter_test.db"
-        with open_connection(db_path) as conn:
-            rebuild_index(conn)
-        now = datetime.now(timezone.utc)
-        (ConversationBuilder(db_path, "conv1")
-         .provider("claude")
-         .created_at(now.isoformat())
-         .save())
-        backend = SQLiteBackend(db_path)
-        return ConversationRepository(backend)
-
-    def test_since_with_datetime_object(self, filter_repo):
-        """Test .since() with datetime object instead of string."""
-        dt = datetime(2024, 1, 1, tzinfo=timezone.utc)
-        filter_obj = ConversationFilter(filter_repo).since(dt)
-        assert filter_obj._since_date == dt
-
-    def test_until_with_datetime_object(self, filter_repo):
-        """Test .until() with datetime object instead of string."""
-        dt = datetime(2024, 12, 31, tzinfo=timezone.utc)
-        filter_obj = ConversationFilter(filter_repo).until(dt)
-        assert filter_obj._until_date == dt
-
-    def test_since_and_until_together_datetime(self, filter_repo):
-        """Test both since/until with datetime objects."""
-        start = datetime(2024, 1, 1, tzinfo=timezone.utc)
-        end = datetime(2024, 12, 31, tzinfo=timezone.utc)
-        filter_obj = ConversationFilter(filter_repo).since(start).until(end)
-        assert filter_obj._since_date == start
-        assert filter_obj._until_date == end
-
-
-class TestFiltersSimilarAndBranches:
-    """Test similar() and branch-related predicates (lines 297-298, 324, 339, 381)."""
-
-    @pytest.fixture
-    def filter_repo(self, tmp_path):
-        db_path = tmp_path / "filter_branches.db"
-        with open_connection(db_path) as conn:
-            rebuild_index(conn)
-
-        # Root conversation
-        (ConversationBuilder(db_path, "root")
-         .provider("claude")
-         .save())
-
-        # Continuation (child with parent)
-        (ConversationBuilder(db_path, "cont")
-         .provider("claude")
-         .parent_conversation("root")
-         .branch_type("continuation")
-         .save())
-
-        # Sidechain
-        (ConversationBuilder(db_path, "side")
-         .provider("claude")
-         .parent_conversation("root")
-         .branch_type("sidechain")
-         .save())
-
-        # With branching messages
-        (ConversationBuilder(db_path, "branched")
-         .provider("claude")
-         .add_message("m1", role="user", text="test")
-         .add_message("m2", role="assistant", text="resp1", branch_index=0)
-         .add_message("m3", role="assistant", text="resp2", branch_index=1)
-         .save())
-
-        backend = SQLiteBackend(db_path)
-        return ConversationRepository(backend)
-
-    def test_similar_text(self, filter_repo):
-        """Test .similar() stores text for vector search."""
-        filter_obj = ConversationFilter(filter_repo).similar("test query")
-        assert filter_obj._similar_text == "test query"
-
-    def test_is_continuation_true(self, filter_repo):
-        """Test .is_continuation(True) filter."""
-        filter_obj = ConversationFilter(filter_repo).is_continuation(True)
-        # Verify predicate was added
-        assert len(filter_obj._predicates) == 1
-
-    def test_is_continuation_false(self, filter_repo):
-        """Test .is_continuation(False) to exclude continuations."""
-        filter_obj = ConversationFilter(filter_repo).is_continuation(False)
-        assert len(filter_obj._predicates) == 1
-
-    def test_is_sidechain_true(self, filter_repo):
-        """Test .is_sidechain(True) filter."""
-        filter_obj = ConversationFilter(filter_repo).is_sidechain(True)
-        assert len(filter_obj._predicates) == 1
-
-    def test_is_sidechain_false(self, filter_repo):
-        """Test .is_sidechain(False) to exclude sidechains."""
-        filter_obj = ConversationFilter(filter_repo).is_sidechain(False)
-        assert len(filter_obj._predicates) == 1
-
-    def test_has_branches_true(self, filter_repo):
-        """Test .has_branches(True) to find conversations with message branches."""
-        filter_obj = ConversationFilter(filter_repo).has_branches(True)
-        assert len(filter_obj._predicates) == 1
-
-    def test_has_branches_false(self, filter_repo):
-        """Test .has_branches(False) to exclude conversations with branches."""
-        filter_obj = ConversationFilter(filter_repo).has_branches(False)
-        assert len(filter_obj._predicates) == 1
-
-
-class TestFiltersApplyFiltersLogic:
-    """Test _apply_filters with all branches (lines 411, 413-414, 452, 498, 519, 521)."""
-
-    @pytest.fixture
-    def filter_repo_populated(self, tmp_path):
-        db_path = tmp_path / "filter_populated.db"
-        with open_connection(db_path) as conn:
-            rebuild_index(conn)
-
-        # Mix of providers with various message lengths
-        (ConversationBuilder(db_path, "claude1")
-         .provider("claude")
-         .title("Short")
-         .add_message("m1", text="a")
-         .save())
-
-        (ConversationBuilder(db_path, "gpt1")
-         .provider("chatgpt")
-         .title("Long conversation")
-         .add_message("m1", text="word " * 50)  # 250 words ≈ 62 tokens
-         .add_message("m2", text="more " * 100)  # 500 words ≈ 125 tokens
-         .save())
-
-        (ConversationBuilder(db_path, "claude2")
-         .provider("claude")
-         .title("Medium")
-         .add_message("m1", text="test " * 10)
-         .add_message("m2", text="data " * 15)
-         .save())
-
-        backend = SQLiteBackend(db_path)
-        return ConversationRepository(backend)
-
-    def test_excluded_providers_filter(self, filter_repo_populated):
-        """Test filtering out specific providers."""
-        results = (ConversationFilter(filter_repo_populated)
-                   .no_provider("chatgpt")
-                   .list())
-        assert all(c.provider != "chatgpt" for c in results)
-
-    def test_sort_by_tokens(self, filter_repo_populated):
-        """Test sorting by token count (line 498)."""
-        results = (ConversationFilter(filter_repo_populated)
-                   .sort("tokens")
-                   .list())
-        # Should be sortable without error
-        assert len(results) >= 0
-
-    def test_sort_by_words(self, filter_repo_populated):
-        """Test sorting by word count (line 519)."""
-        results = (ConversationFilter(filter_repo_populated)
-                   .sort("words")
-                   .list())
-        assert len(results) >= 0
-
-    def test_sort_by_longest(self, filter_repo_populated):
-        """Test sorting by longest message (line 521)."""
-        results = (ConversationFilter(filter_repo_populated)
-                   .sort("longest")
-                   .list())
-        assert len(results) >= 0
-
-    def test_sort_by_messages(self, filter_repo_populated):
-        """Test sorting by message count."""
-        results = (ConversationFilter(filter_repo_populated)
-                   .sort("messages")
-                   .list())
-        assert len(results) >= 0
-
-
-class TestFiltersIDPrefixResolution:
-    """Test ID prefix resolution paths (lines 576-577, 591-592, 606-613)."""
-
-    @pytest.fixture
-    def filter_repo_with_id(self, tmp_path):
-        db_path = tmp_path / "filter_id.db"
-        with open_connection(db_path) as conn:
-            rebuild_index(conn)
-
-        (ConversationBuilder(db_path, "abc123def456")
-         .provider("claude")
-         .save())
-
-        backend = SQLiteBackend(db_path)
-        return ConversationRepository(backend)
-
-    def test_id_prefix_exact_match_fast_path(self, filter_repo_with_id):
-        """Test ID prefix fast path when prefix resolves to single conversation."""
-        # This should use the resolve_id fast path
-        results = (ConversationFilter(filter_repo_with_id)
-                   .id("abc123")
-                   .list())
-        # Expected behavior depends on backend implementation
-        assert isinstance(results, list)
-
-    def test_fts_search_exception_handling(self, filter_repo_with_id):
-        """Test FTS search fallback on exception (line 606-613)."""
-        filter_obj = ConversationFilter(filter_repo_with_id)
-        filter_obj.contains("test")
-        # Mock the search to raise an exception
-        with patch.object(filter_repo_with_id, 'search', side_effect=Exception("FTS error")):
-            # Should fall back to list
-            results = filter_obj.list()
-            assert isinstance(results, list)
-
-
-class TestFiltersListSummariesPaths:
-    """Test list_summaries() and all its branches (lines 669-685, 784-787, 801-802, etc)."""
-
-    @pytest.fixture
-    def filter_repo_summaries(self, tmp_path):
-        db_path = tmp_path / "filter_summaries.db"
-        with open_connection(db_path) as conn:
-            rebuild_index(conn)
-
-        (ConversationBuilder(db_path, "with_summary")
-         .provider("claude")
-         .metadata({"summary": "This is a summary"})
-         .add_message("m1", text="Message")
-         .save())
-
-        (ConversationBuilder(db_path, "without_summary")
-         .provider("claude")
-         .add_message("m1", text="Message")
-         .save())
-
-        backend = SQLiteBackend(db_path)
-        return ConversationRepository(backend)
-
-    def test_list_summaries_with_provider_filter(self, filter_repo_summaries):
-        """Test list_summaries with provider filter."""
-        results = (ConversationFilter(filter_repo_summaries)
-                   .provider("claude")
-                   .list_summaries())
-        assert all(isinstance(s, ConversationSummary) for s in results)
-
-    def test_list_summaries_with_tag_filter(self, filter_repo_summaries):
-        """Test list_summaries with tag filter."""
-        results = (ConversationFilter(filter_repo_summaries)
-                   .tag("mytag")
-                   .list_summaries())
-        assert isinstance(results, list)
-
-    def test_list_summaries_with_summary_has_type(self, filter_repo_summaries):
-        """Test list_summaries with has('summary') filter (line 857)."""
-        results = (ConversationFilter(filter_repo_summaries)
-                   .has("summary")
-                   .list_summaries())
-        # Should include conversations with summaries
-        assert all(s.summary for s in results)
-
-    def test_list_summaries_cannot_use_content_filters(self, filter_repo_summaries):
-        """Test that list_summaries rejects content-dependent filters."""
-        with pytest.raises(ValueError, match="content-dependent filters"):
-            (ConversationFilter(filter_repo_summaries)
-             .has("thinking")
-             .list_summaries())
-
-    def test_can_use_summaries_check(self, filter_repo_summaries):
-        """Test can_use_summaries() method."""
-        # Should be True for simple filters
-        simple = ConversationFilter(filter_repo_summaries).provider("claude")
-        assert simple.can_use_summaries() is True
-
-        # Should be False for content filters
-        with_content = ConversationFilter(filter_repo_summaries).has("thinking")
-        assert with_content.can_use_summaries() is False
-
-
-class TestFiltersPick:
-    """Test pick() interactive picker (lines 725-744, 701)."""
-
-    @pytest.fixture
-    def filter_repo_pick(self, tmp_path):
-        db_path = tmp_path / "filter_pick.db"
-        with open_connection(db_path) as conn:
-            rebuild_index(conn)
-
-        for i in range(5):
-            (ConversationBuilder(db_path, f"conv{i}")
-             .provider("claude")
-             .title(f"Conversation {i}")
-             .save())
-
-        backend = SQLiteBackend(db_path)
-        return ConversationRepository(backend)
-
-    def test_pick_no_results(self, filter_repo_pick):
-        """Test pick() with no matching conversations."""
-        result = (ConversationFilter(filter_repo_pick)
-                  .provider("nonexistent")
-                  .pick())
-        assert result is None
-
-    def test_pick_non_tty_returns_first(self, filter_repo_pick):
-        """Test pick() in non-TTY returns first result without prompting."""
-        with patch("sys.stdout.isatty", return_value=False):
-            result = (ConversationFilter(filter_repo_pick)
-                      .pick())
-            assert result is not None
-
-    def test_pick_empty_input_returns_first(self, filter_repo_pick):
-        """Test pick() with empty input returns first."""
-        with patch("sys.stdout.isatty", return_value=True):
-            with patch("builtins.input", return_value=""):
-                result = (ConversationFilter(filter_repo_pick)
-                          .pick())
-                assert result is not None
-
-    def test_pick_valid_choice(self, filter_repo_pick):
-        """Test pick() with valid numeric choice."""
-        with patch("sys.stdout.isatty", return_value=True):
-            with patch("builtins.input", return_value="1"):
-                result = (ConversationFilter(filter_repo_pick)
-                          .pick())
-                assert result is not None
-
-    def test_pick_invalid_choice_number(self, filter_repo_pick):
-        """Test pick() with out-of-range choice."""
-        with patch("sys.stdout.isatty", return_value=True):
-            with patch("builtins.input", return_value="999"):
-                result = (ConversationFilter(filter_repo_pick)
-                          .pick())
-                assert result is None
-
-    def test_pick_invalid_input_value_error(self, filter_repo_pick):
-        """Test pick() with non-numeric input."""
-        with patch("sys.stdout.isatty", return_value=True):
-            with patch("builtins.input", return_value="not a number"):
-                result = (ConversationFilter(filter_repo_pick)
-                          .pick())
-                assert result is None
-
-    def test_pick_eof_error(self, filter_repo_pick):
-        """Test pick() handling EOFError."""
-        with patch("sys.stdout.isatty", return_value=True):
-            with patch("builtins.input", side_effect=EOFError):
-                result = (ConversationFilter(filter_repo_pick)
-                          .pick())
-                assert result is None
-
-    def test_pick_keyboard_interrupt(self, filter_repo_pick):
-        """Test pick() handling KeyboardInterrupt."""
-        with patch("sys.stdout.isatty", return_value=True):
-            with patch("builtins.input", side_effect=KeyboardInterrupt):
-                result = (ConversationFilter(filter_repo_pick)
-                          .pick())
-                assert result is None
-
-
-class TestFiltersNegativeFTSLogic:
-    """Test negative FTS terms and has_post_filters (lines 591-592, etc)."""
-
-    @pytest.fixture
-    def filter_repo_fts(self, tmp_path):
-        db_path = tmp_path / "filter_fts.db"
-        with open_connection(db_path) as conn:
-            rebuild_index(conn)
-
-        (ConversationBuilder(db_path, "conv1")
-         .provider("claude")
-         .add_message("m1", text="error in the system")
-         .save())
-
-        (ConversationBuilder(db_path, "conv2")
-         .provider("claude")
-         .add_message("m1", text="working perfectly")
-         .save())
-
-        backend = SQLiteBackend(db_path)
-        return ConversationRepository(backend)
-
-    def test_negative_fts_excludes_conversations(self, filter_repo_fts):
-        """Test no_contains() excludes conversations with term."""
-        results = (ConversationFilter(filter_repo_fts)
-                   .no_contains("error")
-                   .list())
-        # Should exclude conv1
-        assert not any("error" in c.display_title or any("error" in (m.text or "").lower() for m in c.messages) for c in results)
-
-
-# =============================================================================
-# UNIFIED.PY TESTS
-# =============================================================================
-
-
-class TestUnifiedMissingRole:
-    """Test _missing_role() error (lines 33-34)."""
-
-    def test_missing_role_raises_error(self):
-        """Test _missing_role() raises ValueError."""
-        from polylogue.schemas.unified import _missing_role
-        with pytest.raises(ValueError, match="Message has no role"):
-            _missing_role()
-
-
-class TestUnifiedExtractReasoningTraces:
-    """Test extract_reasoning_traces with all branches (line 129, 137)."""
-
-    def test_extract_reasoning_traces_empty_content(self):
-        """Test with empty/None content."""
-        result = extract_reasoning_traces(None, "claude")
-        assert result == []
-
-    def test_extract_reasoning_traces_non_dict_block(self):
-        """Test with non-dict block (line 128-129)."""
-        result = extract_reasoning_traces(["string", 123], "claude")
-        assert result == []
-
-    def test_extract_reasoning_traces_thinking_block(self):
-        """Test with thinking block (type='thinking')."""
-        content = [{"type": "thinking", "thinking": "Let me think..."}]
-        result = extract_reasoning_traces(content, "claude")
-        assert len(result) == 1
-        assert result[0].text == "Let me think..."
-
-    def test_extract_reasoning_traces_gemini_thought_block(self):
-        """Test Gemini isThought format (line 136-137)."""
-        content = [{"isThought": True, "text": "Gemini thinking"}]
-        result = extract_reasoning_traces(content, "gemini")
-        assert len(result) == 1
-        assert result[0].text == "Gemini thinking"
-
-    def test_extract_reasoning_traces_thinking_fallback_text(self):
-        """Test thinking block without 'thinking' field falls back to 'text'."""
-        content = [{"type": "thinking", "text": "Fallback text"}]
-        result = extract_reasoning_traces(content, "claude")
-        assert len(result) == 1
-        assert result[0].text == "Fallback text"
-
-
-class TestUnifiedExtractContentBlocks:
-    """Test extract_content_blocks with all block types (lines 159, 189, 224-233, 275, 330, 365)."""
-
-    def test_extract_content_blocks_empty(self):
-        """Test with None content."""
-        result = extract_content_blocks(None)
-        assert result == []
-
-    def test_extract_content_blocks_non_dict_items(self):
-        """Test with non-dict items in content (line 188)."""
-        result = extract_content_blocks(["string", 123, None])
-        assert result == []
-
-    def test_extract_content_blocks_text_block(self):
-        """Test text block extraction."""
-        content = [{"type": "text", "text": "Hello"}]
-        result = extract_content_blocks(content)
-        assert len(result) == 1
-        from polylogue.lib.viewports import ContentType
-        assert result[0].type == ContentType.TEXT
-
-    def test_extract_content_blocks_thinking_block(self):
-        """Test thinking block extraction."""
-        content = [{"type": "thinking", "thinking": "Thought"}]
-        result = extract_content_blocks(content)
-        assert len(result) == 1
-
-    def test_extract_content_blocks_tool_use_block(self):
-        """Test tool_use block extraction."""
-        content = [{
-            "type": "tool_use",
-            "name": "bash",
-            "id": "tool1",
-            "input": {"command": "ls"}
-        }]
-        result = extract_content_blocks(content)
-        assert len(result) == 1
-
-    def test_extract_content_blocks_tool_result_block(self):
-        """Test tool_result block extraction (line 224-231)."""
-        content = [{"type": "tool_result", "content": "result data"}]
-        result = extract_content_blocks(content)
-        assert len(result) == 1
-
-    def test_extract_content_blocks_tool_result_no_content(self):
-        """Test tool_result with missing content field."""
-        content = [{"type": "tool_result"}]
-        result = extract_content_blocks(content)
-        assert len(result) == 1
-        assert result[0].text == ""
-
-    def test_extract_content_blocks_code_block(self):
-        """Test code block extraction (line 232-240)."""
-        content = [{
-            "type": "code",
-            "text": "print('hello')",
-            "language": "python"
-        }]
-        result = extract_content_blocks(content)
-        assert len(result) == 1
-
-    def test_extract_content_blocks_code_block_code_field(self):
-        """Test code block with 'code' field instead of 'text'."""
-        content = [{"type": "code", "code": "def test(): pass"}]
-        result = extract_content_blocks(content)
-        assert len(result) == 1
-
-    def test_extract_content_blocks_default_text_type(self):
-        """Test unknown block type defaults to (or skipped)."""
-        content = [{"type": "unknown", "data": "something"}]
-        result = extract_content_blocks(content)
-        # Unknown types not in the if/elif chain are skipped
-        assert len(result) == 0
-
-
-class TestUnifiedExtractTokenUsage:
-    """Test extract_token_usage edge cases."""
-
-    def test_extract_token_usage_none(self):
-        """Test with None usage."""
-        result = extract_token_usage(None)
-        assert result is None
-
-    def test_extract_token_usage_empty_dict(self):
-        """Test with empty dict returns None (no tokens)."""
-        result = extract_token_usage({})
-        # Empty dict has no tokens, returns None
-        assert result is None
-
-    def test_extract_token_usage_partial_fields(self):
-        """Test with partial token fields."""
-        result = extract_token_usage({"input_tokens": 100})
-        assert result.input_tokens == 100
-
-
-class TestUnifiedExtractTextHelpers:
-    """Test text extraction helpers (lines 275, 330, 365)."""
-
-    def test_extract_claude_code_text_none(self):
-        """Test extract_claude_code_text with None."""
-        result = extract_claude_code_text(None)
-        assert result == ""
-
-    def test_extract_claude_code_text_non_dict_items(self):
-        """Test with non-dict items (line 274)."""
-        result = extract_claude_code_text(["string", 123])
-        assert result == ""
-
-    def test_extract_claude_code_text_text_and_thinking(self):
-        """Test combining text and thinking blocks."""
-        content = [
-            {"type": "text", "text": "Text part"},
-            {"type": "thinking", "thinking": "Thinking part"}
-        ]
-        result = extract_claude_code_text(content)
-        assert "Text part" in result
-        assert "Thinking part" in result
-
-    def test_extract_chatgpt_text_none(self):
-        """Test extract_chatgpt_text with None."""
-        result = extract_chatgpt_text(None)
-        assert result == ""
-
-    def test_extract_chatgpt_text_parts_not_list(self):
-        """Test with parts not being a list (line 290-291)."""
-        result = extract_chatgpt_text({"parts": "string"})
-        assert result == "string"
-
-    def test_extract_chatgpt_text_no_parts(self):
-        """Test with missing parts."""
-        result = extract_chatgpt_text({})
-        assert result == ""
-
-    def test_extract_chatgpt_text_non_string_parts(self):
-        """Test with non-string items in parts (only strings are included)."""
-        result = extract_chatgpt_text({"parts": [123, "text", {"key": "val"}]})
-        # Only string items are included
-        assert result == "text"
-
-    def test_extract_codex_text_not_list(self):
-        """Test extract_codex_text with non-list content (line 297)."""
-        result = extract_codex_text({"data": "not a list"})
-        assert result == ""
-
-    def test_extract_codex_text_multiple_text_fields(self):
-        """Test codex with multiple possible text fields."""
-        content = [
-            {"input_text": "input"},
-            {"output_text": "output"},
-            {"text": "text"}
-        ]
-        result = extract_codex_text(content)
-        assert "input" in result or "output" in result or "text" in result
-
-
-class TestUnifiedExtractHarmonizedMessage:
-    """Test extract_harmonized_message with all providers (line 338)."""
-
-    def test_extract_harmonized_message_invalid_provider(self):
-        """Test with unknown provider (line 338)."""
-        with pytest.raises(ValueError, match="Unknown provider"):
-            extract_harmonized_message("unknown_provider", {})
-
-    def test_extract_harmonized_message_claude_code(self):
-        """Test Claude Code extraction."""
-        raw = {
-            "uuid": "msg1",
-            "timestamp": "2024-01-01T00:00:00Z",
-            "message": {
-                "role": "user",
-                "content": [{"type": "text", "text": "Hello"}],
-                "model": "claude"
-            }
-        }
-        result = extract_harmonized_message("claude-code", raw)
-        assert isinstance(result, HarmonizedMessage)
-
-    def test_extract_harmonized_message_claude_ai(self):
-        """Test Claude AI extraction."""
-        raw = {
-            "uuid": "msg1",
-            "sender": "user",
-            "text": "Hello",
-            "created_at": "2024-01-01T00:00:00Z"
-        }
-        result = extract_harmonized_message("claude-ai", raw)
-        assert isinstance(result, HarmonizedMessage)
-
-    def test_extract_harmonized_message_chatgpt(self):
-        """Test ChatGPT extraction."""
-        raw = {
-            "id": "msg1",
-            "author": {"role": "user"},
-            "content": {"parts": ["Hello"]},
-            "create_time": 1704067200
-        }
-        result = extract_harmonized_message("chatgpt", raw)
-        assert isinstance(result, HarmonizedMessage)
-
-    def test_extract_harmonized_message_gemini(self):
-        """Test Gemini extraction."""
-        raw = {
-            "role": "user",
-            "text": "Hello"
-        }
-        result = extract_harmonized_message("gemini", raw)
-        assert isinstance(result, HarmonizedMessage)
-
-    def test_extract_harmonized_message_codex(self):
-        """Test Codex extraction."""
-        raw = {
-            "id": "msg1",
-            "role": "user",
-            "content": [{"text": "Hello"}],
-            "timestamp": "2024-01-01T00:00:00Z"
-        }
-        result = extract_harmonized_message("codex", raw)
-        assert isinstance(result, HarmonizedMessage)
-
-
-class TestUnifiedHarmonizeParsedMessage:
-    """Test harmonize_parsed_message edge cases (lines 501-509)."""
-
-    def test_harmonize_parsed_message_none_meta(self):
-        """Test with None provider_meta (line 501-502)."""
-        result = harmonize_parsed_message("claude", None)
-        assert result is None
-
-    def test_harmonize_parsed_message_not_message_record(self):
-        """Test with non-message record (line 506-507)."""
-        result = harmonize_parsed_message("claude-code", {"type": "metadata"})
-        assert result is None
-
-    def test_harmonize_parsed_message_valid(self):
-        """Test valid harmonization."""
-        meta = {
-            "raw": {
-                "uuid": "msg1",
-                "sender": "user",
-                "text": "Hello"
-            }
-        }
-        result = harmonize_parsed_message("claude-ai", meta)
-        assert isinstance(result, HarmonizedMessage)
-
-
-class TestUnifiedBulkHarmonize:
-    """Test bulk_harmonize edge cases (lines 525-532)."""
-
-    def test_bulk_harmonize_no_provider_meta(self):
-        """Test with parsed messages without provider_meta (line 527)."""
-        class MockParsedMessage:
-            pass
-
-        messages = [MockParsedMessage()]
-        result = bulk_harmonize("claude", messages)
-        assert result == []
-
-    def test_bulk_harmonize_mixed_valid_invalid(self):
-        """Test with mix of valid and invalid records."""
-        class MockParsedMessage:
-            def __init__(self, meta=None):
-                self.provider_meta = meta
-
-        messages = [
-            MockParsedMessage({"raw": {"type": "metadata"}}),  # Not a message record - for claude-code
-            MockParsedMessage({"raw": {"type": "user", "uuid": "1", "message": {"role": "user", "content": [{"type": "text", "text": "Hi"}]}}}),  # Valid
-        ]
-        result = bulk_harmonize("claude-code", messages)
-        assert len(result) == 1
-
-
-class TestUnifiedIsMessageRecord:
-    """Test is_message_record for Claude Code type checking."""
-
-    def test_is_message_record_claude_code_user(self):
-        """Test Claude Code user message."""
-        result = is_message_record("claude-code", {"type": "user"})
-        assert result is True
-
-    def test_is_message_record_claude_code_assistant(self):
-        """Test Claude Code assistant message."""
-        result = is_message_record("claude-code", {"type": "assistant"})
-        assert result is True
-
-    def test_is_message_record_claude_code_metadata(self):
-        """Test Claude Code metadata record."""
-        result = is_message_record("claude-code", {"type": "metadata"})
-        assert result is False
-
-    def test_is_message_record_other_provider(self):
-        """Test other providers always return True."""
-        result = is_message_record("chatgpt", {"type": "anything"})
-        assert result is True
-
-
-# =============================================================================
-# VALIDATOR.PY TESTS
-# =============================================================================
-
-
-class TestValidatorImportErrorHandling:
-    """Test jsonschema ImportError handling (lines 26-29, 78)."""
-
-    def test_validator_jsonschema_not_installed(self):
-        """Test SchemaValidator when jsonschema is not available."""
-        with patch("polylogue.schemas.validator.jsonschema", None):
-            with pytest.raises(ImportError, match="jsonschema not installed"):
-                SchemaValidator({})
-
-
-class TestValidatorAvailableProviders:
-    """Test available_providers() method (line 78)."""
-
-    def test_available_providers_missing_schema_dir(self):
-        """Test when SCHEMA_DIR doesn't exist."""
-        with patch("polylogue.schemas.validator.SCHEMA_DIR") as mock_dir:
-            mock_dir.exists.return_value = False
-            result = SchemaValidator.available_providers()
-            assert result == []
-
-
-class TestValidatorDetectDrift:
-    """Test drift detection with all branches (lines 112, 170, 175, 182-186)."""
-
-    def test_validate_detects_unexpected_field(self):
-        """Test detecting unexpected fields (line 164-166)."""
-        schema = {
-            "type": "object",
-            "properties": {"name": {"type": "string"}},
-            "additionalProperties": False
-        }
-        validator = SchemaValidator(schema, strict=True)
-        result = validator.validate({"name": "test", "extra": "field"})
-        assert result.has_drift
-        assert any("Unexpected field" in w for w in result.drift_warnings)
-
-    def test_validate_additional_properties_true(self):
-        """Test schema with additionalProperties: true (line 167-169)."""
-        schema = {
-            "type": "object",
-            "properties": {"name": {"type": "string"}},
-            "additionalProperties": True
-        }
-        validator = SchemaValidator(schema, strict=True)
-        result = validator.validate({"name": "test", "extra": "field"})
-        assert not result.has_drift
-
-    def test_validate_additional_properties_schema(self):
-        """Test additionalProperties with schema dict (line 170-175)."""
-        schema = {
-            "type": "object",
-            "properties": {"name": {"type": "string"}},
-            "additionalProperties": {"type": "string"}
-        }
-        validator = SchemaValidator(schema, strict=True)
-        result = validator.validate({"name": "test", "extra": "value"})
-        # Should detect drift for unexpected field
-        assert True  # Logic depends on implementation
-
-    def test_validate_nested_object_drift(self):
-        """Test nested object drift detection (line 179-180)."""
-        schema = {
-            "type": "object",
-            "properties": {
-                "user": {
-                    "type": "object",
-                    "properties": {"name": {"type": "string"}}
-                }
-            }
-        }
-        validator = SchemaValidator(schema, strict=True)
-        data = {"user": {"name": "test", "extra": "field"}}
-        result = validator.validate(data)
-        # May detect nested drift depending on implementation
-        assert isinstance(result, ValidationResult)
-
-    def test_validate_list_items_drift(self):
-        """Test array items drift detection (line 181-186)."""
-        schema = {
-            "type": "object",
-            "properties": {
-                "items": {
-                    "type": "array",
-                    "items": {
-                        "type": "object",
-                        "properties": {"id": {"type": "integer"}}
-                    }
-                }
-            }
-        }
-        validator = SchemaValidator(schema, strict=True)
-        data = {"items": [{"id": 1, "extra": "field"}]}
-        result = validator.validate(data)
-        assert isinstance(result, ValidationResult)
-
-
-class TestValidatorFormatError:
-    """Test _format_error method (line 197)."""
-
-    def test_validate_multiple_errors(self):
-        """Test validation with multiple errors (line 166, 170)."""
-        schema = {
-            "type": "object",
-            "required": ["name"],
-            "properties": {
-                "name": {"type": "string"},
-                "age": {"type": "integer"}
-            }
-        }
-        validator = SchemaValidator(schema, strict=False)
-        result = validator.validate({"age": "not an integer"})
-        assert len(result.errors) >= 1
-
-
-class TestValidatorConvenienceFunction:
-    """Test validate_provider_export convenience function."""
-
-    def test_validate_provider_export_raises_on_missing_schema(self):
-        """Test that invalid provider raises error."""
-        with pytest.raises(FileNotFoundError):
-            validate_provider_export({}, "invalid_provider", strict=True)
-
-
-class TestValidationResult:
-    """Test ValidationResult class methods."""
-
-    def test_validation_result_has_drift_property(self):
-        """Test has_drift property."""
-        result = ValidationResult(
-            is_valid=True,
-            drift_warnings=["Field X is new"]
-        )
-        assert result.has_drift is True
-
-    def test_validation_result_no_drift(self):
-        """Test has_drift when empty."""
-        result = ValidationResult(is_valid=True)
-        assert result.has_drift is False
-
-    def test_validation_result_raise_if_invalid(self):
-        """Test raise_if_invalid method."""
-        result = ValidationResult(
-            is_valid=False,
-            errors=["Error 1", "Error 2"]
-        )
-        with pytest.raises(ValueError, match="Schema validation failed"):
-            result.raise_if_invalid()
-
-    def test_validation_result_raise_if_valid(self):
-        """Test raise_if_invalid when valid."""
-        result = ValidationResult(is_valid=True)
-        # Should not raise
-        result.raise_if_invalid()
