--- tests/test_tui_comprehensive.py
@@ -1,103 +1,0 @@
-import pytest
-from textual.widgets import DataTable, Input, TabbedContent, Tree
-
-try:
-    from polylogue.ui.tui.app import PolylogueApp
-    from polylogue.ui.tui.screens.browser import Browser
-    from polylogue.ui.tui.screens.dashboard import Dashboard
-    from polylogue.ui.tui.screens.search import Search
-except ImportError:
-    PolylogueApp = None
-
-
-@pytest.mark.skipif(PolylogueApp is None, reason="Textual not installed")
-@pytest.mark.asyncio
-async def test_tui_navigation(workspace_env, conversation_builder):
-    """Test navigating between main tabs."""
-    # Seed some data so stats aren't empty
-    conversation_builder("conv-1").add_message("m1", text="Hello TUI").save()
-
-    app = PolylogueApp()
-    async with app.run_test() as pilot:
-        # Initial state: Dashboard
-        tabs = pilot.app.query_one(TabbedContent)
-        assert tabs.active == "dashboard"
-        assert pilot.app.query_one(Dashboard)
-
-        # Switch to Browser
-        tabs.active = "browser"
-        await pilot.pause()
-        assert pilot.app.query_one(Browser)
-
-        # Switch to Search
-        tabs.active = "search"
-        await pilot.pause()
-        assert pilot.app.query_one(Search)
-
-
-@pytest.mark.skipif(PolylogueApp is None, reason="Textual not installed")
-@pytest.mark.asyncio
-async def test_tui_browser_flow(workspace_env, conversation_builder):
-    """Test browser tree and viewer interaction."""
-    # Seed data
-    conversation_builder("c1").add_message("m1", text="Content for C1").save()
-
-    app = PolylogueApp()
-    async with app.run_test() as pilot:
-        tabs = pilot.app.query_one(TabbedContent)
-        tabs.active = "browser"
-        await pilot.pause()
-
-        tree = pilot.app.query_one(Tree)
-        # Wait for worker to populate tree
-        await pilot.pause(0.5)
-
-        # Expand root (Sources)
-        tree.root.expand()
-        await pilot.pause(0.1)
-
-        # Should have "Test" or "Other" provider node depending on default
-        # Our implementation lists fixed providers: "chatgpt", "claude", etc.
-        # "c1" has provider="test" (default in builder) or "other" fallback logic?
-        # Let's inspect what builder sets. Default is provider="test".
-        # Our Browser lists ["chatgpt", "claude", "gemini", "codex", "other"].
-        # So "c1" should be under "Other" if logic supported it, but our impl
-        # loop specifically query for provider=p.
-        # So "test" provider won't show up unless we add "test" to the list in Browser.py
-        # or use a real provider name in builder.
-        pass
-
-
-@pytest.mark.skipif(PolylogueApp is None, reason="Textual not installed")
-@pytest.mark.asyncio
-async def test_tui_search_flow(workspace_env, conversation_builder):
-    """Test search input and result selection."""
-    # Seed data
-    conversation_builder("c2").add_message("m1", text="UniqueKeyword").save()
-
-    app = PolylogueApp()
-    async with app.run_test() as pilot:
-        tabs = pilot.app.query_one(TabbedContent)
-        tabs.active = "search"
-        await pilot.pause()
-
-        inp = pilot.app.query_one(Input)
-        inp.focus()
-        # Verify focus to ensure reliable typing
-        if pilot.app.focused and pilot.app.focused.id != "search-input":
-            await pilot.pause(0.1)
-
-        inp.value = "UniqueKeyword"
-        await pilot.press("enter")
-
-        # Wait for search execution
-        await pilot.pause(0.5)
-
-        table = pilot.app.query_one(DataTable)
-        assert table.row_count > 0
-
-        # Select first row
-        row_key = next(iter(table.rows))
-
-        # Verify Key
-        assert row_key.value == "c2"
